requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pymupdf>=1.23.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0
aiofiles>=23.0.0
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0 